Merge manual.md into readme.md
(Based on user feedback)

Visiting the main page of the repository, one would expect to see the
contents of the docs in the README page/front page directly. Storing the
list of manuals in a separate file does not bring a lot of benefit
compared to just listing them in the README and having direct links to
the example projects.

Also improved the structure and titles of some of the docs and added
a one-sentence introduction to each section.

Signed-off-by: Violin Yanev <violinyanev@users.noreply.github.com>