Improve documentation structure
We already had some structure in the main README which clusted the
docs by a few categories. We added more stuff over time, but never
updated the structure of the pages. This makes it harder to find stuff,
and also the list of folders in the root directory is a bit too large
and pushes the main page further down so that a visitor won't see it.

This PR doesn't change the docs themselves, but restructures them a bit.
There are three main folders:
/basics
/advanced
/debugging

Fixed the links too. Also, added some new lines in some of the docs
which had too long lines.